Add imports and exports for character and common vms
