+++ charactersheet/charactersheet/viewmodels/character/armor/index.js
@@ -1,9 +1,15 @@
 'use strict';
 
 import ko from 'knockout'
-//TODO Import
 
-function ArmorViewModel() {
+import { Armor} from 'charactersheet/common/models'
+import { CharacterManager } from 'charactersheet/utilities'
+import { Notifications } from 'charactersheet/utilities'
+import { PersistenceService } from 'charactersheet/services/common'
+
+import template from './index.html'
+
+export function ArmorViewModel() {
     var self = this;
 
     self.blankArmor = ko.observable(new Armor());
@@ -191,3 +197,8 @@ function ArmorViewModel() {
         });
     };
 }
+
+ko.components.register('armor', {
+  viewModel: ArmorViewModel,
+  template: template
+})
