+++ charactersheet/charactersheet/viewmodels/character/skills_tab/index.js
@@ -1,9 +1,19 @@
 'use strict';
 
+import ko from 'knockout'
+
+import { FeatsViewModel } from 'charactersheet/viewmodels/character/feats'
+import { FeaturesViewModel } from 'charactersheet/viewmodels/character/features'
+import { ProficienciesViewModel } from 'charactersheet/viewmodels/character/proficiencies'
+import { SkillsViewModel } from 'charactersheet/viewmodels/character/skills'
+import { TrackerViewModel } from 'charactersheet/viewmodels/character/tracker'
+import { TraitsViewModel } from 'charactersheet/viewmodels/character/traits'
+import { ViewModelUtilities } from 'charactersheet/utilities'
+
 /**
  * This view model contains the player's skills information.
  */
-function SkillsTabViewModel() {
+export function SkillsTabViewModel() {
     var self = this;
 
     self.skillsViewModel       = ko.observable(new SkillsViewModel());
