+++ charactersheet/charactersheet/viewmodels/character/spells_tab/index.js
@@ -1,9 +1,17 @@
 'use strict';
 
+import ko from 'knockout'
+
+import { ActionsToolbarViewModel } from 'charactersheet/viewmodels/character/actions_toolbar'
+import { SpellbookViewModel } from 'charactersheet/viewmodels/character/spells'
+import { SpellSlotsViewModel } from 'charactersheet/viewmodels/character/spell_slots'
+import { SpellStatsViewModel } from 'charactersheet/viewmodels/character/spell_stats'
+import { ViewModelUtilities } from 'charactersheet/utilities'
+
 /**
  * This view model contains the player's spells information.
  */
-function SpellsTabViewModel() {
+export function SpellsTabViewModel() {
     var self = this;
 
     self.actionsToolbarViewModel = ko.observable(new ActionsToolbarViewModel());
