+++ charactersheet/charactersheet/viewmodels/character/traits/index.js
@@ -1,6 +1,18 @@
 'use strict';
 
-function TraitsViewModel() {
+import ko from 'knockout'
+
+import { CharacterManager } from 'charactersheet/utilities'
+import { DataRepository } from 'charactersheet/utilities'
+import { Notifications } from 'charactersheet/utilities'
+import { PersistenceService } from 'charactersheet/services/common'
+import { SortService } from 'charactersheet/services/common'
+import { Tracked } from 'charactersheet/character/models'
+import { Trait } from 'charactersheet/character/models'
+
+import template from './index.html'
+
+export function TraitsViewModel() {
     var self = this;
 
     self.DESCRIPTION_MAX_LENGTH = 45;
@@ -187,3 +199,8 @@ function TraitsViewModel() {
         return 'Tracked Traits are listed in the Tracker.';
     };
 }
+
+ko.components.register('traits', {
+  viewModel: TraitsViewModel,
+  template: template
+})
