+++ charactersheet/charactersheet/viewmodels/common/chat_tab/index.js
@@ -1,6 +1,13 @@
 'use strict';
 
-function ChatTabViewModel() {
+import ko from 'knockout'
+
+import { CharacterManager } from 'charactersheet/utilities'
+import { Notifications } from 'charactersheet/utilities'
+import { PersistenceService } from 'charactersheet/services/common'
+import { ViewModelUtilities } from 'charactersheet/utilities'
+
+export function ChatTabViewModel() {
     var self = this;
 
     self.chatViewModel = ko.observable(new ChatViewModel());
