+++ charactersheet/charactersheet/viewmodels/common/index.js
@@ -1,0 +1,25 @@
+import ko from 'knockout'
+
+import './character_picker'
+import './characters'
+import './chat/chat_cell'
+import './chat/chat_detail'
+import './chat/chat_log_chat_item'
+import './chat/chat_log_image_item'
+import './chat/chat_log_read_aloud_item'
+import './chat/chat_log_system_item'
+import './chat/chat_modal'
+import './chat_tab'
+import './login'
+import './master_detail'
+import './notes'
+import './notes_tab'
+import './notifications'
+import './party_manager'
+import './player_image'
+import './player_info'
+import './root'
+import './user_notification'
+import './wizard/steps'
+
+ko.applyBindings()
