+++ charactersheet/charactersheet/viewmodels/common/player_image/index.js
@@ -1,6 +1,15 @@
 'use strict';
 
-function PlayerImageViewModel() {
+import ko from 'knockout'
+
+import { CharacterManager } from 'charactersheet/utilities'
+import { Fixtures } from 'charactersheet/utilities'
+import { Notifications } from 'charactersheet/utilities'
+import { PersistenceService } from 'charactersheet/services/common'
+
+import template from './index.html'
+
+export function PlayerImageViewModel() {
     var self = this;
 
     self.openModal = ko.observable(false);
@@ -193,3 +202,8 @@ function PlayerImageViewModel() {
         }
     };
 }
+
+ko.components.register('party-image', {
+  viewModel: PlayerImageViewModel,
+  template: template
+})
